Replace JDK 8 API
`Math.toIntExact` is a JDK 8 API. Not all Android apps want/need to include JDK 8, since it comes at a non-zero cost. In this case, `Long.intValue` should behave identically since only small Long values should exist. There doesn't seem to be any other JDK 8 APIs in this plugin.

+++ packages/camera/camera_android_camerax/android/src/main/java/io/flutter/plugins/camerax/CameraSelectorHostApiImpl.java
@@ -32,7 +32,7 @@ public class CameraSelectorHostApiImpl implements CameraSelectorHostApi {
     CameraSelector cameraSelector;
 
     if (lensFacing != null) {
-      cameraSelector = cameraSelectorBuilder.requireLensFacing(Math.toIntExact(lensFacing)).build();
+      cameraSelector = cameraSelectorBuilder.requireLensFacing(lensFacing.intValue()).build();
     } else {
       cameraSelector = cameraSelectorBuilder.build();
     }
